Creación dinámica de las rutas

+++ app/routes/Model.js
@@ -4,21 +4,25 @@ module.exports = function(modelName, extras = []) {
 
     const controllerCallbacks = require(`../controllers/${modelName}Controller`)
     
-    extras.push({ method: 'GET', endpoint: "", callback: 'index'});
-    extras.push({ method: 'GET', endpoint: ":id", callback: 'show'});
-    extras.push({ method: 'POST', endpoint: "", callback: 'store'});
-    extras.push({ method: 'PUT', endpoint: ":id", callback: 'update'});
-    extras.push({ method: 'PATCH', endpoint: ":id", callback: 'update'});
-    extras.push({ method: 'DELETE', endpoint: ":id", callback: 'destroy'});
+    if(extras.length == 0)
+    {
+        extras.push({ method: 'GET', endpoint: "", callback: 'index'});
+        extras.push({ method: 'GET', endpoint: ":id", callback: 'show'});
+        extras.push({ method: 'POST', endpoint: "", callback: 'store'});
+        extras.push({ method: 'PUT', endpoint: ":id", callback: 'update'});
+        extras.push({ method: 'PATCH', endpoint: ":id", callback: 'update'});
+        extras.push({ method: 'DELETE', endpoint: ":id", callback: 'destroy'});
+    }
 
     if (extras.length > 0)
     {
+        
         extras.forEach(route => {
             var fn = route.callback.trim();
             if (fn in controllerCallbacks && typeof controllerCallbacks[fn] === "function") 
             {
                 const routePath = `/${modelName.toLowerCase()}/${route.endpoint}`
-                console.log(route.method.toLowerCase() + ": " +routePath)
+                console.log(`${route.method.toUpperCase()}: ${routePath}   (${fn})`)
 
 
                 if(route.method.toLowerCase() === "get")
